Backend/odc/requirements.txt
altair==5.3.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
CORS-Analyzer==1.0.9
Django==5.2
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psycopg2-binary==2.9.10
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
sqlparse==0.5.3
streamlit==1.35.0
tenacity==8.4.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.13.0
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
gunicorn